Added if statement to prevent more than 50 messages being purged at a time in the same channel.

+++ bot.py
@@ -254,18 +254,23 @@ async def self(interaction: discord.Interaction, user: discord.User):
 @tree.command(name="purge", description="Slash command for Purging a channel.")
 async def self(interaction: discord.Interaction, channel: discord.TextChannel, number: int):
     try:
-        modrole = discord.utils.get(interaction.guild.roles, id=getmodrole(interaction.guild.id)[0])
-        if modrole in interaction.user.roles:
-            await interaction.response.defer(ephemeral=True)
-            messages = channel.history(limit=number)
-            async for a in messages:
-                await a.delete()
-                await asyncio.sleep(3)
-            await interaction.followup.send(f"Deleted {number} message(s)")
-        else:
+        if number > 50:
             await interaction.response.send_message(
-                content=f"""You don't have proper permissions to run this command.""",
+                content=f"""Cannot purge more than 50 messages at a time.""",
                 ephemeral=True)
+        else:
+            modrole = discord.utils.get(interaction.guild.roles, id=getmodrole(interaction.guild.id)[0])
+            if modrole in interaction.user.roles:
+                await interaction.response.defer(ephemeral=True)
+                messages = channel.history(limit=number)
+                async for a in messages:
+                    await a.delete()
+                    await asyncio.sleep(3)
+                await interaction.followup.send(f"Deleted {number} message(s)")
+            else:
+                await interaction.response.send_message(
+                    content=f"""You don't have proper permissions to run this command.""",
+                    ephemeral=True)
     except Exception as e:
         print(e)
         await interaction.response.send_message(content=f"""Something went wrong.""", ephemeral=True)
